Split off pillboxMenus
