+++ src/components/index.ts
@@ -3,6 +3,5 @@ export { default as InlineMathInputView } from "components/InlineMathInputView";
 export { default as SegmentedControl } from "components/SegmentedControl";
 export { default as StaticMathQuillView } from "components/StaticMathQuillView";
 export { default as Button } from "components/Button";
-export { default as PillboxContainer } from "components/PillboxContainer";
 export { default as Toggle } from "components/Toggle";
 export { default as IconButton } from "components/IconButton";
