+++ src/plugins/performance-info/PerformanceView.tsx
@@ -22,11 +22,15 @@ export class PerformanceView extends Component<{
               <IconButton
                 iconClass={"dsm-icon-bookmark"}
                 onTap={() => {
-                  this.props.mainController().toggleMenuPinned();
+                  this.props
+                    .mainController()
+                    .enabledPlugins.pillboxMenus?.toggleMenuPinned();
                 }}
                 btnClass={() => ({
                   "dsm-pi-pin-menu-button": true,
-                  "dsm-selected": this.props.mainController().pillboxMenuPinned,
+                  "dsm-selected":
+                    this.props.mainController().enabledPlugins.pillboxMenus
+                      ?.pillboxMenuPinned,
                 })}
               />
             </Tooltip>
