+++ src/plugins/performance-info/index.ts
@@ -1,20 +1,29 @@
+import { jquery } from "../../utils/depUtils";
 import Controller from "./Controller";
-import { destroyView, initView } from "./View";
+import { MainPopupFunc } from "./PerformanceView";
 import { Plugin } from "plugins";
 
-export let controller: Controller;
+let controller: Controller;
 
 const performanceInfo: Plugin = {
   id: "performance-info",
   key: "performanceInfo",
   onEnable: (c) => {
     controller = new Controller(c);
-    initView(c);
-    return undefined;
+    const pm = c.enabledPlugins.pillboxMenus;
+    pm?.addPillboxButton({
+      id: "dsm-pi-menu",
+      tooltip: "performance-info-name",
+      iconClass: "dsm-icon-pie-chart",
+      popup: () => MainPopupFunc(controller, c),
+    });
+    return controller;
   },
   onDisable: (c) => {
     controller.stop();
-    destroyView(c);
+    const pm = c.enabledPlugins.pillboxMenus;
+    pm?.removePillboxButton("dsm-pi-menu");
+    jquery(document).off(".expanded-menu-view");
   },
   enabledByDefault: false,
 } as const;
