+++ src/plugins/pillbox-menus/components/PillboxContainer.less
