+++ src/plugins/pillbox-menus/components/PillboxMenu.less
