+++ src/preload/moduleOverrides/pillbox.replacements
@@ -21,7 +21,7 @@ __from__
 $DCGView.Components.IfElse(
     () => window.DesModder,
     {
-      true: () => window.DesModder?.controller.pillboxButtonsView(true),
+      true: () => DSM.pillboxMenus?.pillboxButtonsView(true),
       false: null
     }
 ),
@@ -43,7 +43,7 @@ $DCGView.createElement($, {
 $DCGView.Components.IfElse(
     () => window.DesModder,
     {
-      true: () => window.DesModder?.controller.pillboxButtonsView(false),
+      true: () => DSM.pillboxMenus?.pillboxButtonsView(false),
       false: null
     }
 ), __from__
@@ -60,7 +60,7 @@ this.controller.isGraphSettingsOpen() ? `bottom: ${$e}px;` : "bottom: auto"
 
 *Replace* `from` with
 ```js
-this.controller.isGraphSettingsOpen() || window.DesModder?.controller?.isSomePillboxMenuOpen()
+this.controller.isGraphSettingsOpen() || DSM.pillboxMenus?.isSomePillboxMenuOpen()
   ? `bottom: ${$e}px;` : "bottom: auto;"
 ```
 
@@ -84,7 +84,7 @@ __from__,
 $DCGView.Components.IfElse(
     () => window.DesModder,
     {
-      true: () => window.DesModder?.controller.pillboxMenuView(true),
+      true: () => DSM.pillboxMenus?.pillboxMenuView(true),
       false: null
     }
 )
@@ -140,5 +140,5 @@ shouldShowContainer() {
 
 *Replace* `ret` with
 ```js
-__ret__ || (DesModder?.controller && !DesModder?.controller.showHorizontalPillboxMenu())
+__ret__ || (!DSM.pillboxMenus?.showHorizontalPillboxMenu())
 ```
